[order] delivery 이벤트 스키마 변경
